requirements.txt
fastapi==0.109.2
gunicorn==21.2.0
uvicorn[standard]==0.27.0.post1
transformers==4.37.2
torch==2.2.0
torchaudio== 2.2.0
kaldiio==2.18.0
espnet==202402
scipy==1.12.0